pyproject.toml
[project]
name = "olist-seller-success-dashboard"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.117.1",
    "lightgbm>=4.6.0",
    "matplotlib>=3.10.6",
    "mlxtend>=0.23.4",
    "nltk>=3.9.1",
    "pandas>=2.3.2",
    "prophet>=1.1.7",
    "psycopg2-binary>=2.9.10",
    "python-dotenv>=1.1.1",
    "scikit-learn>=1.7.2",
    "seaborn>=0.13.2",
    "torch>=2.8.0",
    "transformers>=4.56.1",
    "uvicorn>=0.36.0",
]
